+++ test/e2e/app-dir/metadata-edge/app/another/opengraph-image.tsx
@@ -1,0 +1,5 @@
+export const alt = 'Open Graph Edge'
+
+export const runtime = 'edge'
+
+export { default } from '../og'
