+++ test/e2e/app-dir/metadata-edge/app/another/page.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function Page() {
+  return <>hello another</>
+}
+
+export const metadata = {
+  title: 'another page',
+}
+
+export const runtime = 'edge'
