+++ test/e2e/app-dir/metadata-edge/app/og.tsx
@@ -1,0 +1,21 @@
+import { ImageResponse } from 'next/server'
+
+export default function og() {
+  return new ImageResponse(
+    (
+      <div
+        style={{
+          width: '100%',
+          height: '100%',
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'center',
+          fontSize: 128,
+          background: 'lavender',
+        }}
+      >
+        Open Graph
+      </div>
+    )
+  )
+}
