+++ test/e2e/app-dir/metadata-edge/app/opengraph-image.tsx
@@ -1,23 +1,3 @@
-import { ImageResponse } from 'next/server'
-
 export const alt = 'Open Graph'
 
-export default function og() {
-  return new ImageResponse(
-    (
-      <div
-        style={{
-          width: '100%',
-          height: '100%',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          fontSize: 128,
-          background: 'lavender',
-        }}
-      >
-        Open Graph
-      </div>
-    )
-  )
-}
+export { default } from './og'
